Nu mai visezi angle.cpp

+++ MC stuff/types/angle.cpp
@@ -2,7 +2,7 @@
 
 Angle::Angle() : v(0) { }
 Angle::Angle(byte value) : v(value) { }
-Angle::Angle(double value) : v(fmod(value, 360) / 360 * 256) { }
+Angle::Angle(double value) : v(byte(fmod(value, 360) / 360 * 256)) { }
 
 Angle::operator double()
 {
@@ -10,5 +10,5 @@ Angle::operator double()
 }
 void Angle::operator=(double value)
 {
-	v = (fmod(value, 360) / 360 * 256);
+	v = byte((fmod(value, 360) / 360 * 256));
 }
